+++ unpack/GFA.py
@@ -4,7 +4,7 @@ from util.filesystem import *
 from util.utils import byterepr
 import util.rawutil as rawutil
 
-GFAC_META_STRUCT = '4s11I/11[4I]/11[n]128a'
+GFAC_META_STRUCT = '4s11I /11[4I] /11[n] 128a'
 GFCP_HEADER_STRUCT = '4s4I'
 GFAC_HASH_MULTIPLIER = 0x89
 
@@ -20,7 +20,8 @@ class extractGFA (rawutil.TypeReader):
 		self.outdir = make_outdir(filename)
 		ptr = self.readmeta(data)
 		self.readinfo()
-		self.gfcp = data[ptr:]
+		self.data = data
+		self.gfcp_offset = data.tell()
 	
 	def calc_hash(self, name):
 		hash = 0
@@ -31,7 +32,7 @@ class extractGFA (rawutil.TypeReader):
 		return hash
 	
 	def readmeta(self, data):
-		hdata, ptr = self.unpack_from(GFAC_META_STRUCT, data, 0, getptr=True)
+		hdata = self.unpack_from(GFAC_META_STRUCT, data, 0)
 		magic = hdata[0]
 		if magic != b'GFAC':
 			error.InvalidMagicError('Invalid magic %s, expected GFAC' % byterepr(magic))
@@ -48,7 +49,6 @@ class extractGFA (rawutil.TypeReader):
 		if self.verbose:
 			print('Version: %08x' % self.version)
 			print('File count: %d' % self.filecount)
-		return ptr
 	
 	def readinfo(self):
 		self.nodes = [self.makenode(self.fat[i], self.filenames[i][0]) for i in range(0, self.filecount)]
@@ -66,9 +66,8 @@ class extractGFA (rawutil.TypeReader):
 		return node
 	
 	def extract(self):
-		data = self.gfcp
-		magic, version, comp, self.decsize, compsize = self.unpack_from(GFCP_HEADER_STRUCT, data)
-		data = data[20:]
+		magic, version, comp, self.decsize, compsize = self.unpack_from(GFCP_HEADER_STRUCT, self.data)
+		data = self.data.read()
 		if magic != b'GFCP':
 			error.InvalidMagicError('Invalid GFCP magic: %s' % byterepr(magic))
 		if comp in (2, 3):
@@ -78,15 +77,16 @@ class extractGFA (rawutil.TypeReader):
 			if self.verbose:
 				print('Decompressed')
 		elif comp == 1:
-			error.UnsupportedCompressionError('Actually unsupported compression')
+			error.UnsupportedCompressionError('Currently unsupported compression')
 		else:
 			error.UnsupportedCompressionError('Invalid compression')
 		for node in self.nodes:
 			outname = self.outdir + node.name
 			if self.verbose:
 				print('Extracting %s' % outname)
-			filedata = data[node.offset:node.offset + node.length]
-			bwrite(filedata, outname)
+			#data.seek(dataoffset + node.offset)
+			with open(outname, 'wb') as file:
+				file.write(data[node.offset: node.offset + node.length])
 	
 	def decompressLZ10(self, data):
 		# cannot use compress.LZ10 because of strange header